requirements.txt
Pillow~=8.1.0
tqdm~=4.45.0
yacs~=0.1.8
vizer~=0.1.5
tensorboardX
pycocotools
numpy~=1.19.5
torch~=1.4.0
setuptools~=54.0.0
six~=1.15.0
opencv-python~=4.2.0.34
torchvision~=0.5.0